commiting

+++ domain/users/user_dto.go
@@ -12,13 +12,14 @@ import (
 
 //User : User struct for performing some operations
 type User struct {
-	ID          int    `json:"id"`
+	ID          int64  `json:"id"`
 	FirstName   string `json:"first_name"`
 	LastName    string `json:"last_name"`
 	Email       string `json:"email"`
 	DateCreated string `json:"date_created"`
 }
 
+//Validate : it filter the email
 func (user *User) Validate() *errors.RestErr {
 	user.Email = strings.TrimSpace(strings.ToLower(user.Email))
 	if user.Email == "" {
